pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dodal"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
description = "Ophyd devices and other utils that could be used across DLS beamlines"
dependencies = [
    "ophyd",
    "bluesky",
    "pyepics",
    "dataclasses-json",
    "pillow",
    "requests",
    "graypy",
    "pydantic",
]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.rst"
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "databroker",
    "area_detector_handlers",
    "black",
    "mypy",
    "flake8-isort",
    "Flake8-pyproject",
    "mockito",
    "pipdeptree",
    "pre-commit",
    "pydata-sphinx-theme>=0.12",
    "pytest-cov",
    "pytest-random-order",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "sphinx-design",
    "tox-direct",
    "types-requests",
    "types-mock",
]

[project.urls]
GitHub = "https://github.com/DiamondLightSource/dodal"

[[project.authors]] # Further authors may be added by duplicating this section
email = "dominic.oram@diamond.ac.uk"
name = "Dominic Oram"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dodal = ["*.txt"]

[tool.setuptools_scm]
write_to = "src/dodal/_version.py"

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true # Ignore missing stubs in imported modules

[tool.isort]
float_to_top = true
profile = "black"

[tool.flake8]
extend-ignore = [
    "E203", # See https://github.com/PyCQA/pycodestyle/issues/373
    "F811", # support typing.overload decorator
    "F722", # allow Annotated[typ, some_func("some string")]
    "E501",
    "W503",
]
max-line-length = 88 # Respect black's line length (default 88),
exclude = [".tox", "venv"]


[tool.pytest.ini_options]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
markers = [
    "s03: marks tests as requiring the s03 simulator running (deselect with '-m \"not s03\"')",
    "adsim: marks tests as requiring the simulated area detector simulator running (deselect with '-m \"not adsim\"')",
]
addopts = """
    --tb=native -vv --doctest-modules --doctest-glob="*.rst"
    --cov=dodal --cov-report term --cov-report xml:cov.xml
    """
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "docs src tests"

[tool.coverage.run]
data_file = "/tmp/dodal.coverage"

[tool.coverage.paths]
# Tests are run from installed location, map back to the src directory
source = ["src", "**/site-packages/"]

# tox must currently be configured via an embedded ini string
# See: https://github.com/tox-dev/tox/issues/999
[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist=True

[testenv:{pre-commit,mypy,pytest,docs}]
# Don't create a virtualenv for the command, requires tox-direct plugin
direct = True
passenv = *
allowlist_externals =
    pytest
    pre-commit
    mypy
    sphinx-build
    sphinx-autobuild
commands =
    pytest: pytest -m 'not s03' {posargs}
    mypy: mypy src tests --ignore-missing-imports --no-strict-optional {posargs}
    pre-commit: pre-commit run --all-files {posargs}
    docs: sphinx-{posargs:build -EW --keep-going} -T docs build/html
"""
